check-examples: compile against the release, not against main
A reader installs a RELEASE, so an example is correct when it compiles against
the release - and main is ahead of it by design. The gap is not theoretical:
z2ui5_cl_ui5_view_builder sat on main from 2026-08-12 while the newest tag was
1.142.0 from 2026-07-20, so for three weeks a page could teach the view
builder, pass this gate, and not work for a single reader.

The release comes from the three places check-version already holds honest,
which is why it can be trusted here; both gates now read them through
scripts/lib/release.mjs rather than carrying two copies of the pattern.
A2UI5_REF overrides it - that is how an against-main canary runs, ahead of a
release rather than instead of it. The run prints which framework it compiled
against, so it cannot claim more than it checked.

Verified: all 46 examples compile and lint clean against 1.143.0.

Also adds release-freshness: check:version fails when the site names a release
that is not the newest, but it only ran on pull requests and pushes - so a
release published in a quiet week went unnoticed. That is precisely the
incident it was written for (1.143.0 published at 12:20, the site said 1.142.0
for the rest of the day). Weekly now, as an issue rather than a red badge on a
branch nobody pushed.
